+++ appcenter/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## [0.6.0](https://github.com/capacitor-community/appcenter-sdk-capacitor/compare/@capacitor-community/appcenter@0.5.0...@capacitor-community/appcenter@0.6.0)
+
+### Features
+
+* **android**: Add `AppCenter.setNetworkRequestsAllowed()` API to block any network requests without disabling the SDK.
+
+### Chores
+
+* **refactor**: `AppCenter.setNetworkRequestsAllowed`
+
 ## 0.5.0
 
 ### Features